# uid lookup cache