@fw42 2015-04-25
Hm I think this might not be thread-safe