@fd0 2015-04-26
No, that's not thread-safe. Btw: you can compile restic with `go build -race` to enable the built-in race detector.