@fw42 2015-04-26
Pushed a mutex. Could use some guidance how to profile this. I used pprof before, but I'm not sure if that's the right tool here.