@fw42 2015-04-26
> you can compile restic with `go build -race` to enable the built-in race detector

maybe that should be part of the test suite and CI run?