@fd0 2015-05-01
I did some benchmarks: `BenchmarkArchiveDirectory` reports a 13.5% decrease in runtime for archiving a directory with 50k very small files (2 byte random each) in it:

    $ go test -v -run x -bench BenchmarkArchiveDirectory -test.benchdir ~/tmp/testdir -benchmem -benchtime 30s -cpu 4
    [...]

    $ benchcmp /tmp/{before,after}
    benchmark                       old ns/op      new ns/op      delta
    BenchmarkArchiveDirectory-4     5447585912     4712772814     -13.49%

    benchmark                       old allocs     new allocs     delta
    BenchmarkArchiveDirectory-4     8767356        8518639        -2.84%

    benchmark                       old bytes     new bytes     delta
    BenchmarkArchiveDirectory-4     942622028     928083420     -1.54%

However, `BenchmarkNodeFillExtra` shows that the overall runtime for `fillExtra()` is really small already, but has been reduced drastically by the cache:

    benchmark                 old ns/op     new ns/op     delta
    BenchmarkNodeFillUser     14987         1007          -93.28%

    benchmark                 old allocs     new allocs     delta
    BenchmarkNodeFillUser     11             6              -45.45%

    benchmark                 old bytes     new bytes     delta
    BenchmarkNodeFillUser     193           80            -58.55%

I'm unable to see any great improvements in real-world backups, but I'd like to keep the cache nonetheless. It's not that much code to care about.